+++ packages/core/src/markView/CoreMarkView.ts
@@ -3,7 +3,7 @@ import type { Mark } from 'prosemirror-model'
 import type { EditorView, MarkView, ViewMutationRecord } from 'prosemirror-view'
 
 import { createKey } from '../create-key'
-import { isContentDOMRemoval } from '../is-content-dom-removal'
+import { isContentDOMRemoval } from '../utils/is-content-dom-removal'
 
 import type { CoreMarkViewSpec, CoreMarkViewUserOptions, MarkViewDOMSpec } from './CoreMarkViewOptions'
