+++ packages/core/src/utils/is-content-dom-removal.ts
